instance_manager: Move out Runner to its own module
